DB 저장까지 확인 (수정)
